Conflicts:
	aslam_cv/include/aslam/cameras/camera.h
	aslam_cv/include/aslam/cameras/distortion.h
	aslam_cv/src/distortion.cc

+++ aslam_cv/include/aslam/cameras/distortion-equidistant.h
@@ -125,7 +125,7 @@ class EquidistantDistortion : public aslam::Distortion {
 
   /// \brief Returns the number of parameters used in the distortion model.
   ///        NOTE: Use the constexpr function parameterCount if you know the exact distortion type.
-  virtual size_t getParameterSize() const {
+  virtual int getParameterSize() const {
     return kNumOfParams;
   }
